+++ ModelDeploy/models/_ONNXSmoke.py
@@ -1,45 +1,27 @@
 # -*- coding: utf-8 -*-
 
 from typing import Any, Dict, List, Optional, Tuple
+import onnxruntime as onnxrt
 import numpy as np
 import time
 import cv2
-import ctypes
-
 import torch
 from torch import Tensor
 from torch.nn import functional as F
-
-import pycuda.driver as cuda
-import pycuda.autoinit
-import tensorrt as trt
-
+from . import ModelBase
 from .. import modules as md
 
-# TensorRT logger singleton
-TRT_LOGGER = trt.Logger(trt.Logger.WARNING)
-EXPLICIT_BATCH = 1 << (int)(trt.NetworkDefinitionCreationFlag.EXPLICIT_BATCH)
+
+__all__ = ['ONNXSmoke']
 
 
-class HostDeviceMem(object):
-    def __init__(self, host_mem, device_mem):
-        self.host = host_mem
-        self.device = device_mem
-
-    def __str__(self):
-        return "Host:\n" + str(self.host) + "\nDevice:\n" + str(self.device)
-
-    def __repr__(self):
-        return self.__str__()
-
-
-class TRTSmoke:
+class ONNXSmoke(ModelBase):
     """
-    원본 : ONNXSmoke Class
-     - 작성자 : 김형석
-    수정 : TRTSmoke
-     - 수정자 : 한상준
-     - 수정 내용 : ONNX → TensorRT
+    원본 : SmokeInfer Class
+     - 작성자 : 한상준
+    수정 : ONNXSmoke
+     - 수정자 : 김형석
+     - 수정 내용 : Inference Engine에 적용하기 위해 일부 구조 변경 / 고정된 metadata 삭제, forward 입력 파라미터 변경
 
     """
 
@@ -47,94 +29,42 @@ class TRTSmoke:
                  weight_path:str,
                  input_width:int = 1280,
                  input_height:int = 384,
-                 shared_library_path:str='./ModelDeploy/lib/libmmdeploy/libmmdeploy_trt_net.so'
+                 shared_library_path:str='./ModelDeploy/lib/libmmdeploy/libmmdeploy_ort_net.so',
+                 onnx_providers:Optional[List[str]]=None,
                  ):
-
-        self.__weight_path = weight_path
-        self.__input_width = input_width
-        self.__input_height = input_height
-        self.__input_converter = md.InputConverter(input_width=input_width,
-                                                   input_height=input_height,
-                                                   input_type='ndarray')
-        self.shared_library_path = shared_library_path
+        
+        super().__init__(weight_path, input_width, input_height, 'ndarray')
+        if onnx_providers is None:
+            onnx_providers = ['CUDAExecutionProvider', 'CPUExecutionProvider']
         self.bbox_code_size = 7
         self.bbox_coder = md.SMOKECoder(base_depth=(28.01, 16.32),
                                      base_dims=((0.88, 1.73, 0.67), (1.78, 1.70, 0.58), (3.88, 1.63, 1.53)),
                                      code_size=7)
-
-        ### Load TensorRT Model
-        ctypes.CDLL(self.shared_library_path)
-        trt.init_libnvinfer_plugins(TRT_LOGGER, '')
-        self.trt_runtime = trt.Runtime(TRT_LOGGER)
-
-        with open(self.__weight_path, 'rb') as f:
-            trt_model = f.read()
-        self.engine = self.trt_runtime.deserialize_cuda_engine(trt_model)
-        self.context = self.engine.create_execution_context()
-
-        ### Memory allocation
-        self.inputs = []
-        self.outputs = []
-        self.allocations = []
-        for i in range(self.engine.num_bindings):
-            is_input = False
-            if self.engine.get_tensor_mode(self.engine.get_tensor_name(i)) == trt.TensorIOMode.INPUT:
-                is_input = True
-            name = self.engine.get_tensor_name(i)
-            dtype = self.engine.get_tensor_dtype(name)
-            shape = self.engine.get_tensor_shape(name)
-            if is_input:
-                self.batch_size = shape[0]
-            size = np.dtype(trt.nptype(dtype)).itemsize
-            for s in shape:
-                size *= s
-            allocation = cuda.mem_alloc(size)
-            binding = {
-                "index": i,
-                "name": name,
-                "dtype": np.dtype(trt.nptype(dtype)),
-                "shape": list(shape),
-                "allocation": allocation,
-            }
-            self.allocations.append(allocation)
-            if is_input:
-                self.inputs.append(binding)
-            else:
-                self.outputs.append(binding)
-
-    def cls_score_spec(self):
-        return self.outputs[0]["shape"], self.outputs[0]["dtype"]
-
-    def bbox_pred_spec(self):
-        return self.outputs[1]["shape"], self.outputs[1]["dtype"]
+        session_option = onnxrt.SessionOptions()
+        session_option.register_custom_ops_library(shared_library_path)
+        self.session = onnxrt.InferenceSession(self.__weight_path, sess_options=session_option,
+                                               providers=onnx_providers)
 
     def warmup(self):
-        for idx in range(4):
+        for idx in range(20):
             inputs = np.random.rand(1, 3, self.__input_height, self.__input_width).astype('f')
             start = time.time()
-            cls_score = np.zeros(*self.cls_score_spec())
-            bbox_pred = np.zeros(*self.bbox_pred_spec())
-            cuda.memcpy_htod(self.inputs[0]["allocation"], np.ascontiguousarray(np.random.rand(1, 3, 384, 1280)))
-            self.context.execute_v2(self.allocations)
-            cuda.memcpy_dtoh(cls_score, self.outputs[0]["allocation"])
-            cuda.memcpy_dtoh(bbox_pred, self.outputs[1]["allocation"])
+            _ = self.session.run(None, {"img": inputs})
             print(f"Iter {idx}: {time.time()- start:.5f}sec")
         print("WarmUp completed!")
 
-    def forward(self, image:np.ndarray, meta_data:List[Dict[str, Any]]) -> md.InferenceResult:
+    def _forward(self, image:np.ndarray, meta_data:List[Dict[str, Any]]) -> md.InferenceResult:
         input_data = self.__input_converter(image)
         start = time.time()
+        data = {
+            "img": input_data  # convert image dimention to (1,C,H,W)
+        }
 
         # Inference
-        cls_score = np.zeros(*self.cls_score_spec())
-        bbox_pred = np.zeros(*self.bbox_pred_spec())
-        cuda.memcpy_htod(self.inputs[0]["allocation"], np.ascontiguousarray(input_data))
-        self.context.execute_v2(self.allocations)
-        cuda.memcpy_dtoh(cls_score, self.outputs[0]["allocation"])
-        cuda.memcpy_dtoh(bbox_pred, self.outputs[1]["allocation"])
+        outputs = self.session.run(None, data)
         print(f"Session: {time.time() - start:.5f}sec")
 
-        results = self.predict_by_feat(Tensor(cls_score), Tensor(bbox_pred), meta_data)
+        results = self.predict_by_feat(Tensor(outputs[0]), Tensor(outputs[1]), meta_data)
         result = results[0]
         print("result = ", result)
         scores:torch.Tensor = result['scores_3d']
@@ -184,6 +114,7 @@ class TRTSmoke:
                 scores_3d=scores
             )
             result_list.append(results)
+
         return result_list
 
     def _decode_heatmap(self,
@@ -195,13 +126,13 @@ class TRTSmoke:
                         kernel: int = 3) -> Tuple[Tensor, Tensor, Tensor]:
         bs, _, feat_h, feat_w = cls_score.shape
 
-        center_heatmap_pred = TRTSmoke.get_local_maximum(cls_score, kernel=kernel)
+        center_heatmap_pred = ONNXSmoke.get_local_maximum(cls_score, kernel=kernel)
 
-        *batch_dets, topk_ys, topk_xs = TRTSmoke.get_topk_from_heatmap(
+        *batch_dets, topk_ys, topk_xs = ONNXSmoke.get_topk_from_heatmap(
             center_heatmap_pred, k=topk)
         batch_scores, batch_index, batch_topk_labels = batch_dets
 
-        regression = TRTSmoke.transpose_and_gather_feat(reg_pred, batch_index)
+        regression = ONNXSmoke.transpose_and_gather_feat(reg_pred, batch_index)
         regression = regression.view(-1, 8)
 
         points = torch.cat([topk_xs.view(-1, 1),
@@ -235,7 +166,7 @@ class TRTSmoke:
     def transpose_and_gather_feat(feat, ind):
         feat = feat.permute(0, 2, 3, 1).contiguous()
         feat = feat.view(feat.size(0), -1, feat.size(3))
-        feat = TRTSmoke.gather_feat(feat, ind)
+        feat = ONNXSmoke.gather_feat(feat, ind)
         return feat
 
     @staticmethod
